% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aqdm.R
\name{get_aqdm}
\alias{get_aqdm}
\title{get_aqdm}
\usage{
get_aqdm(endpoint, params)
}
\arguments{
\item{endpoint}{the name of an AQS REST service endpoint}

\item{params}{a named list of parameters to be passed to the REST service}
}
\description{
generic function for making requests to the AQS REST Service
}
\details{
All the other data retrieval functions in this package are just
wrappers for this function with additional validation
}

