% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_list}
\alias{get_list}
\title{Make a list
request}
\usage{
get_list(name, ...)
}
\arguments{
\item{name}{The variable for which to show a list of valid values}

\item{\dots}{Qualifier variables to filter the list on.}
}
\description{
Returns a list of valid values for a variable. Requesting a list takes three
parts. The list name, the qualifier (what to filter the list on), and the
resource - only one of which is currently available.
}
\details{
Must include a \code{name} parameter telling the REST service which parameter
list to return. Acceptable values are:
\itemize{
 \item \code{state}
 \item \code{county}
 \item \code{site}
 \item \code{pc}
 \item \code{param}
 \item \code{format}
 \item \code{cbsa}
 \item \code{csa}
 \item \code{duration}
 \item \code{agency}
}
May also pass one or more qualifier parameters. These will filter the
\code{name} list returned. Acceptable qualifier parameters are:
\itemize{
 \item \code{state}
 \item \code{county}
 \item \code{pc}
 \item \code{agency_code}
}
Also accepts a \code{resource} parameter which indicates the name of the
service to use as a reference for this list. Currently, this parameter is
optional and the only acceptable value is \code{rawData}. This value is
inserted into the query by default.
}

