% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rawData.R
\name{get_rawData}
\alias{get_rawData}
\title{Make a rawData
request}
\usage{
get_rawData(...)
}
\arguments{
\item{dots}{parameters to pass to the request query string (see Details)}
}
\description{
Returns raw data that match your query criteria. This is a synchronous query
that returns data to your screen, so you must stay connected to the service
while the query runs.
}

