% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rawDataNotify.R
\name{get_rawDataNotify}
\alias{get_rawDataNotify}
\title{Make a rawDataNotify
request}
\usage{
get_rawDataNotify(...)
}
\arguments{
\item{dots}{parameters to pass to the request query string (see Details)}
}
\description{
Returns raw data that match your query criteria. This is an asynchronous
query that returns a Transaction ID number. The query is run on our servers
and when it is complete, an email will be sent you with a link to your data
(tied to the Transaction ID).
}

