% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retrieve.R
\name{get_retrieve}
\alias{get_retrieve}
\title{Make a retrieve
request}
\usage{
get_retrieve(x)
}
\arguments{
\item{x}{either an \code{aqdmObject} object or a transaction id}
}
\description{
Returns the data created with a previsously submitted request.
}

