% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_serviceAvailable.R
\name{get_serviceAvailable}
\alias{get_serviceAvailable}
\title{Make a serviceAvailable
request}
\usage{
get_serviceAvailable()
}
\description{
Returns a response if the REST services are working.
}
\details{
Will return READY @" followed by the date and time at the server (US Eastern
Time Zone) if the service is running, otherwise will return an error.
}

