% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Make a status
request}
\usage{
get_status(x)
}
\arguments{
\item{x}{either an \code{aqdmObject} object or a transaction id}
}
\description{
Returns the status of a previously submitted job.
}

